refactor: cleanup controller

+++ src/services/collection.js
@@ -1,17 +1,26 @@
 const {
+  sequelize,
   Collection,
   CollectionRecommendation,
   Recommendation,
 } = require("../model");
-const { BadRequestError } = require("../utils/error");
+const { BadRequestError, NotFoundError } = require("../utils/error");
+const RecommendationService = require("./recommendation");
 
 class CollectionService {
+  constructor() {
+    this.Collection = Collection;
+    this.CollectionRecommendation = CollectionRecommendation;
+    this.Recommendation = Recommendation;
+    this.RecommendationService = RecommendationService;
+  }
+
   create = async ({ name, description, userId }) => {
     try {
       if (!name) {
         throw BadRequestError("Name is required");
       }
-      return await Collection.create({
+      return await this.Collection.create({
         name,
         description,
         user_id: userId,
@@ -24,7 +33,7 @@ class CollectionService {
 
   getById = async ({ collectionId, userId }) => {
     try {
-      return await Collection.findOne({
+      return await this.Collection.findOne({
         where: { id: collectionId, ...(userId && { user_id: userId }) },
       });
     } catch (error) {
@@ -35,14 +44,14 @@ class CollectionService {
 
   getWithRecommendations = async ({ collectionId, userId }) => {
     try {
-      return await Collection.findOne({
+      return await this.Collection.findOne({
         where: {
           id: collectionId,
           user_id: userId,
         },
         include: [
           {
-            model: Recommendation,
+            model: this.Recommendation,
             through: { attributes: [] }, // To exclude the join table attributes
           },
         ],
@@ -55,7 +64,7 @@ class CollectionService {
 
   getAllWithPagination = async ({ userId, limit, page }) => {
     try {
-      return await Collection.findAndCountAll({
+      return await this.Collection.findAndCountAll({
         where: {
           user_id: userId,
         },
@@ -68,10 +77,26 @@ class CollectionService {
     }
   };
 
-  addRecommendation = async (collection, recommendation) => {
+  addRecommendation = async ({ collectionId, recommendationId, userId }) => {
     try {
+      const collection = await this.getById({
+        collectionId,
+        userId: userId,
+      });
+      if (!collection) {
+        throw NotFoundError("Collection not found");
+      }
+
+      const recommendation = await this.RecommendationService.getById({
+        recommendationId,
+        userId: userId,
+      });
+      if (!recommendation) {
+        throw NotFoundError("Recommendation not found");
+      }
+
       await collection.addRecommendation(recommendation, {
-        through: CollectionRecommendation,
+        through: this.CollectionRecommendation,
       });
     } catch (error) {
       console.error(error);
@@ -79,11 +104,52 @@ class CollectionService {
     }
   };
 
-  delete = async (collection) => {
+  removeRecommendation = async ({ collectionId, recommendationId, userId }) => {
     try {
+      const collection = await this.getById({
+        collectionId,
+        userId: userId,
+      });
+      if (!collection) {
+        throw NotFoundError("Collection not found");
+      }
+
+      const recommendation = await this.RecommendationService.getById({
+        recommendationId,
+      });
+      if (!recommendation) {
+        throw NotFoundError("Recommendation not found");
+      }
+
+      const collectionRecommendation = await this.getCollectionRecommendation({
+        collectionId: collection.id,
+        recommendationId: recommendationId,
+      });
+      if (!collectionRecommendation) {
+        throw NotFoundError("Recommendation not found in collection");
+      }
+
+      await collectionRecommendation.destroy();
+    } catch (error) {
+      console.error(error);
+      throw error;
+    }
+  };
+
+  delete = async ({ collectionId, userId }) => {
+    try {
+      const collection = await this.getById({
+        collectionId,
+        userId: userId,
+      });
+
+      if (!collection) {
+        throw NotFoundError("Collection not found");
+      }
+
       await sequelize.transaction(async (t) => {
         await collection.destroy({ transaction: t });
-        await CollectionRecommendation.destroy({
+        await this.CollectionRecommendation.destroy({
           where: { collection_id: collection.id },
           transaction: t,
         });
@@ -96,7 +162,7 @@ class CollectionService {
 
   getCollectionRecommendation = async ({ collectionId, recommendationId }) => {
     try {
-      return await CollectionRecommendation.findOne({
+      return await this.CollectionRecommendation.findOne({
         where: {
           collection_id: collectionId,
           recommendation_id: recommendationId,
